+++ packages/agent/src/subagents/__init__.py
@@ -1,0 +1,1 @@
+# Subagent definitions will be added in Tasks 6-8